Reformatting physics review script.
